refactor(architecture): move all entities into domain/models
