+++ scripts/drop_database.py
