+++ scripts/generate_token.py
