+++ website/init/scheduler.py
@@ -4,7 +4,7 @@ from .extensions import scheduler, db
 
 def schedule_jobs(app):
     def cleanup_expired_codes():
-        from website.models import VerificationCode
+        from website.domain.models import VerificationCode
 
         with app.app_context():
             VerificationCode.delete_expired()
