--- website/models/__init__.py
@@ -1,3 +1,0 @@
-from .user import User, VerificationCode
-from .post import Post, Image, SavedPost, PostImage, Comment
-from .enums import UserRole, UserTheme
